adisulaksono_120140038_pertemuan6/course_management/requirements.txt
arrow==1.3.0
binaryornot==0.4.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
hupper==1.12.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
PasteDeploy==3.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
Pygments==2.19.1
pyramid==2.0.2
pyramid-mako==1.1.0
pyramid_debugtoolbar==4.12.1
pyramid_jinja2==2.10.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
six==1.17.0
text-unidecode==1.3
translationstring==1.4
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
urllib3==2.4.0
venusian==3.1.1
waitress==3.0.2
WebOb==1.8.9
zope.deprecation==5.1
zope.interface==7.2
